refactor: reputation modal to drawer
